Style sheet and other layout issue updates. also renamed download-links module to download-data.
